feat(*): Change project name: angular-google-maps
Angular 4 is coming this year and this project aims to support
all upcoming angular versions.

Because of this reason, we rename the project.

BREAKING CHANGES

** IMPORTANT **
The project and NPM package name changes:

Old name: angular2-google-maps
Old NPM pkg name: angular2-google-maps

New name: angular-google-maps
New NPM pkg name: @angular-google-maps/core

To have a consisten naming pattern, we also change the
component/directives names

| Old class name / old selector | New class name / new selector |
|-------------------------------------------------------------|---------------------------------------|
| SebmGoogleMap / sebm-google-map | AgmMap / agm-map |
| SebmGoogleMapCircle / sebm-google-map-circle | AgmCircle / agm-circle
|
| SebmGoogleMapInfoWindow / sebm-google-map-info-window | AgmInfoWindow
/ agm-info-window |
| SebmGoogleMapKmlLayer / sebm-google-map-kml-layer | AgmKmlLayer /
agm-kml-layer |
| SebmGoogleMapMarker / sebm-google-map-marker | AgmMarker / agm-marker
|
| SebmGoogleMapPolygon / sebm-map-polygon | AgmPolygon / agm-polygon |
| SebmGoogleMapPolyline / sebm-google-map-polyline | AgmPolyline /
agm-polyline |
| SebmGoogleMapPolylinePoint / sebm-google-map-polyline-point |
AgmPolylinePoint / agm-polyline-point |

The package structure also changes. The root directory now contains
**esm** code (ES5 code with ES2015 modules).

A CommonJS compatible format and TS are not part of the package any
more! This change aligns with structure @angular packages.

The styling of the maps is now more intuitive.

To style the map, please use the `agm-map` element (or a custom class or
ID) directly.

The `.sebm-google-map-container` css class is deprecated and will be
gone with the next version.

The `.sebm.google-map-container-inner` is now called
`.agm-map-container-inner`. `.sebm-google-map-container-inner` is
depcrecated and will be gone with the next version.